First commit; Before trying sequelize-cli
