+++ lib/Edge.js
@@ -1,7 +1,7 @@
 /**
  * An Edge in the graph.
  * @typedef {Object} Edge
- * @name module:lib/model.Edge
+ * @name module:lib.Edge
  * @property {number} id - auto-incrementing, primary-key
  * @property {Object} metadata
  * @property {Vertex} from
