+++ lib/index.js
@@ -1,1 +1,28 @@
-const Model = require('./model');
+/**
+ * Model for a simple graph. Initializes the DB and exports the models.
+ * @module lib
+ */
+const Sequelize  = require('sequelize');
+const { config } = require('simian-server-common');
+
+const connectionStringKey = typeof it === 'function' ? 'db.testConnectionString' : 'db.connectionString';
+const connectionString = config.get(connectionStringKey);
+if (!connectionString) {
+  throw new Error(`No value defined for connection string >${connectionStringKey}<`);
+}
+const sequelize        = new Sequelize(connectionString);
+
+const db = {};
+const EdgeType = db['EdgeType'] = sequelize.import('./EdgeType.js');
+const Edge     = db['Edge']     = sequelize.import('./Edge.js');
+const Vertex   = db['Vertex']   = sequelize.import('./Vertex.js');
+
+Edge.belongsTo(EdgeType);
+Edge.belongsTo(Vertex, {as: 'from'});
+Edge.belongsTo(Vertex, {as: 'to'});
+
+db.sequelize = sequelize;
+db.Sequelize = Sequelize;
+
+// sequelize.sync({ force: true });
+module.exports = db;
