+++ package.json
@@ -7,7 +7,7 @@
     "document": "jsdoc -c jsdoc.conf",
     "fix-lint": "eslint lib/ --fix",
     "lint": "eslint lib/",
-    "plato": "plato -d plato -r -e .eslintrc.plato.json lib/",
+    "plato": "plato -d plato -r -e .eslintrc.plato.json ./lib/",
     "test": "node_modules/.bin/nyc node_modules/.bin/mocha test/**/*.js"
   },
   "repository": {
@@ -47,13 +47,13 @@
     "mocha": "^3.5.0",
     "pg": "^7.3.0",
     "pg-hstore": "^2.3.2",
-    "sequelize": "^4.8.0",
+    "sequelize": "^4.10.2",
     "simian-server-common": "0.0.2"
   },
   "devDependencies": {
     "eslint": "^4.6.1",
     "eslint-config-simian": "0.0.3",
-    "jsdoc": "^3.5.4",
+    "jsdoc": "^3.5.5",
     "jsdoc-strip-async-await": "^0.1.0",
     "mockery": "^2.1.0",
     "nyc": "^11.2.0",
